Move CableClub_OnResume to Nurse to Remove PokeCenter Entrance Lag
Entering a Pokemon Center took a long time to load due to the game calling CableClub_OnResume to handle whether or not the nurse should check the Union Room for other players. This check has been moved to the nurse herself, as with the snappier warps the lag entering the Pokemon Center was very apparent. The check is hidden in the wait for the FLDEFF_POKECENTER_HEAL (this is the animation for putting the balls on the machine) so it is imperceptible to the player.

Moved the entire CableClub_OnResume script to C to make it faster as well. The function mimics the functionality of the map script in that it only checks when the map is resumed. This means that the check is only called when the map is refreshed and won't happen multiple times in a row.

+++ include/overworld.h
@@ -84,7 +84,7 @@ void SetDynamicWarpWithCoords(s32 unused, s8 mapGroup, s8 mapNum, s8 warpId, s8 
 void SetWarpDestinationToDynamicWarp(u8 unused);
 void SetWarpDestinationToHealLocation(u8 a1);
 void SetWarpDestinationToLastHealLocation(void);
-void SetLastHealLocationWarp(u8 healLocationId);
+void SetLastHealLocationWarp(u32 healLocationId);
 void UpdateEscapeWarp(s16 x, s16 y);
 void SetEscapeWarp(s8 mapGroup, s8 mapNum, s8 warpId, s8 x, s8 y);
 void SetWarpDestinationToEscapeWarp(void);
